requirements.txt
# Core dependencies for PDF parsing and text processing
PyPDF2>=3.0.0
sentence-transformers>=2.2.0

# Vector database
pinecone>=2.2.0

# Data processing and utilities
numpy>=1.21.0
pathlib2>=2.3.0

# Logging and configuration
python-dotenv>=0.19.0

# PyTorch (required for sentence-transformers)
torch>=2.1.0

# Hugging Face Hub (required for sentence-transformers)
huggingface_hub>=0.16.0